+++ src/main/java/com/lgcms/consulting/api/EmbeddingController.java
@@ -1,6 +1,6 @@
 package com.lgcms.consulting.api;
 
-import com.lgcms.consulting.domain.vectorDB.service.VectorDBService;
+import com.lgcms.consulting.vectorDB.service.VectorDBService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
